Update to clap 4.0
A lot of changes, moved to a derived APIS.. Things should be the same...
Same crappy parser code, but that's going to be a work for another day.

The parsing looks nice now with the derive APIs instead of the weird
copy pastas, get sh*** done fast as i was doing before.

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>