feat: add timezone support for timestamp conversion
Added support for specifying a timezone using the `--timezone` flag or the
`SNAZY_TIMEZONE` environment variable. This allows timestamps to be displayed
in the specified timezone instead of the server's default timezone (usually
UTC). Updated the `Config` struct and related functions to handle timezone
conversion. Added tests to verify the correct parsing and conversion of
timestamps with timezones.

Fixes #276

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>